fix ordering and acceptance test

+++ team_planet_express/src/main/java/shop/core/services/item_list/PagingService.java
@@ -17,7 +17,7 @@ public class PagingService {
         if (pagingEnabled) {
             if (pagingRule != null) {
                 Integer pageSize = Integer.valueOf(pagingRule.getPageSize());
-                long itemCountToSkip = (pagingRule.getPageNumber() - 1) * pageSize;
+                long itemCountToSkip = (long) (pagingRule.getPageNumber() - 1) * pageSize;
                 items = items.stream()
                         .skip(itemCountToSkip)
                         .limit(pageSize)
